# specs/001-nutrition-recipe-planner/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Cookfully API
  version: 0.2.0
  description: >-
    Canonical HTTP contract for recipes, honest nutrition estimates, goals, meal plans,
    grocery lists, exports, agent access, and later goal-aware suggestions. Public decimal
    values are canonical strings. Estimated nutrition and suggestion projections are planning
    aids, not medical advice. MCP tools compose the same application commands and queries.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
  - bearerToken: []
...
paths:
...
  /intelligence/infer:
    post:
      operationId: inferIntelligence
      tags: [Intelligence]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IntelligenceInferenceRequest" }
      responses:
        "200":
          description: A schema-constrained proposal from the local model service.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IntelligenceInferenceResponse" }
  /intelligence/drafts:
    post:
      operationId: createIntelligenceDraft
      tags: [Intelligence]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IntelligenceInferenceRequest" }
      responses:
        "200":
          description: A short-lived reviewed proposal.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IntelligenceDraftResponse" }
  /intelligence/drafts/{draftId}/execute:
    post:
      operationId: executeIntelligenceDraft
      tags: [Intelligence]
      parameters:
        - name: draftId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IntelligenceExecuteRequest" }
      responses:
        "200":
          description: The reviewed actions were executed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IntelligenceExecutionResponse" }
  /intelligence/drafts/{draftId}:
    get:
      operationId: getIntelligenceDraft
      tags: [Intelligence]
      parameters:
        - name: draftId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: The owner-scoped extraction or command proposal.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IntelligenceDraftDetailResponse" }
  /intelligence/extraction-jobs:
    post:
      operationId: createIntelligenceExtractionJob
      tags: [Intelligence]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IntelligenceInferenceRequest" }
      responses:
        "202":
          description: A durable backend-coordinated extraction job was accepted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobAccepted" }
        "422": { $ref: "#/components/responses/ValidationError" }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cookfully_session
    bearerToken:
      type: http
      scheme: bearer
...
  responses:
...
    ValidationError:
      description: One or more fields are invalid.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    IntelligenceInferenceRequest:
      type: object
      required: [operation, prompt]
      additionalProperties: false
      properties:
        operation:
          type: string
          enum: [command, recipe_extract, pantry_extract, cook]
        prompt:
          type: string
          minLength: 1
          maxLength: 50000
        context:
          type: object
          additionalProperties: { type: string }
    IntelligenceInferenceResponse:
      type: object
      required: [status, functionCalls]
      properties:
        status:
          type: string
          enum: [ok, unsupported, unavailable]
        model:
          type: [string, "null"]
        confidence:
          type: [number, "null"]
          minimum: 0
          maximum: 1
        reasoning:
          type: [string, "null"]
        functionCalls:
          type: array
          items:
            type: object
            required: [name, arguments]
            properties:
              name: { type: string }
              arguments: { type: object }
        errorCode:
          type: [string, "null"]
    IntelligenceDraftResponse:
      allOf:
        - $ref: "#/components/schemas/IntelligenceInferenceResponse"
        - type: object
          properties:
            draftId:
              type: [string, "null"]
              format: uuid
            expiresAt:
              type: [string, "null"]
              format: date-time
    IntelligenceExecuteRequest:
      type: object
      required: [confirm]
      additionalProperties: false
      properties:
        confirm: { type: boolean }
    IntelligenceExecutionResponse:
      type: object
      required: [draftId, status, results]
      properties:
        draftId: { type: string, format: uuid }
        status: { type: string, enum: [executed] }
        results:
          type: array
          items: { type: object }
    IntelligenceDraftDetailResponse:
      type: object
      required: [draftId, operation, status, functionCalls, expiresAt]
      properties:
        draftId: { type: string, format: uuid }
        operation: { type: string, enum: [command, recipe_extract, pantry_extract, cook] }
        status: { type: string, enum: [queued, processing, review, executed, expired, failed, unsupported] }
        model: { type: [string, "null"] }
        confidence: { type: [number, "null"], minimum: 0, maximum: 1 }
        reasoning: { type: [string, "null"] }
        functionCalls:
          type: array
          items:
            type: object
            required: [name, arguments]
            properties:
              name: { type: string }
              arguments: { type: object }
        expiresAt: { type: string, format: date-time }
        failureCode: { type: [string, "null"] }
        failureMessage: { type: [string, "null"] }
...
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        fieldErrors:
          type: array
          items:
            type: object
            required: [field, code]
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
...
    JobAccepted:
      type: object
      required: [jobId, status]
      properties:
        jobId: { type: string, format: uuid }
        resourceId: { type: [string, "null"], format: uuid }
        status: { type: string, const: queued }
        coverStatus: { type: [string, "null"], enum: [attached, not_selected, failed, null] }
...